scraper/requirements.txt
feedparser
beautifulsoup4
requests
lxml
spacy
gender-guesser
rapidfuzz
python-dotenv
